requirements.txt
openai>=1.30.0
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
yfinance>=0.2.40
tavily-python>=0.3.0
notion-client>=2.2.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
gTTS>=2.3.0
edge-tts>=6.1.0
